First implementation of the ActionPerformer for pocolibs + code organization
The actions/ directory now contains one file per action (like 'give'). The main
function implementing this action must be decorated with @action.

This action is expected to return a list of atomic action to perform in order
to achieve the action.

A helper function 'genom_request' has been added to build pocolibs request.