Completely revamped the position management
Introduced a new 'PoseManager' that groups all pose-related functions.

PoseManager.get(<smthg>) tries to retreive the pose of <smthg> by querying
ROS, SPARK or a static library of places.
It also normalizes its input to a pose represented as a dict {x,y,z,qx,qy,qz,qw}.

While here: added unit-tests to check the PoseManager works as expected.
Not completely tested (must be tested on a robot with ROS and SPARK).