Added a 'basictake' and a 'take' action
'basictake' simply get the robot to hand its gripper and wait for an object.

'take' plans a move toward the human (with OTP) and waits for an object.

TODO: currently, in 'take', the holder of the object is hardcoded to
'HERAKLES_HUMAN1' because the dialogue does not tell us who holds the object.